Bug 1660127 - Add periodic and manual synchronisation to CardDAV. r=mkmelin
Differential Revision: https://phabricator.services.mozilla.com/D87688

--HG--
rename : mailnews/addrbook/test/unit/test_cardDAV_sync.js => mailnews/addrbook/test/unit/test_cardDAV_syncV1.js
extra : moz-landing-system : lando

+++ mailnews/addrbook/jsaddrbook/AddrBookDirectory.jsm
@@ -106,7 +106,11 @@ function openConnectionTo(file) {
  * Closes the SQLite connection to `file` and removes it from the cache.
  */
 function closeConnectionTo(file) {
-  directories.delete(file.leafName);
+  let directory = directories.get(file.leafName);
+  if (directory) {
+    directory.destroy();
+    directories.delete(file.leafName);
+  }
   let connection = connections.get(file.path);
   if (connection) {
     return new Promise(resolve => {
@@ -170,6 +174,7 @@ class AddrBookDirectory {
     this._fileName = fileName;
     directories.set(fileName, this);
   }
+  destroy() {}
 
   get _prefBranch() {
     if (this.__prefBranch) {
